// web-ecommerce/types/product.ts
export type Product = {
  title: string;
  reviews: number;
  price: number;
  discountedPrice: number;
  id: number;
  imgs?: {
    thumbnails: string[];
    previews: string[];
  };
  category?: string;
  variations?: {
    weight: string;
    price: number;
    discountedPrice: number;
  }[];
};